Move relevant tests tests from frame/value_tests to serialize/row_tests
